volgen: Removed unused Volgen code
Renamed `volgen2` to `volgen`

Updates: #835
