+++ glusterd2/commands/volumes/volfiles.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	restutils "github.com/gluster/glusterd2/glusterd2/servers/rest/utils"
-	volgen "github.com/gluster/glusterd2/glusterd2/volgen2"
+	"github.com/gluster/glusterd2/glusterd2/volgen"
 
 	"github.com/gorilla/mux"
 )
