+++ glusterd2/commands/volumes/volume-delete.go
@@ -7,34 +7,11 @@ import (
 	"github.com/gluster/glusterd2/glusterd2/gdctx"
 	restutils "github.com/gluster/glusterd2/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/glusterd2/transaction"
-	"github.com/gluster/glusterd2/glusterd2/volgen"
 	"github.com/gluster/glusterd2/glusterd2/volume"
 
 	"github.com/gorilla/mux"
 	"github.com/pborman/uuid"
 )
-
-func deleteVolfiles(c transaction.TxnCtx) error {
-
-	var volinfo volume.Volinfo
-	if err := c.Get("volinfo", &volinfo); err != nil {
-		return err
-	}
-
-	if err := volgen.DeleteClientVolfile(&volinfo); err != nil {
-		c.Logger().WithError(err).WithField(
-			"volume", volinfo.Name).Warn("failed to delete client volfile")
-	}
-
-	for _, b := range volinfo.GetLocalBricks() {
-		if err := volgen.DeleteBrickVolfile(&b); err != nil {
-			c.Logger().WithError(err).WithField(
-				"brick", b.Path).Warn("failed to delete brick volfile")
-		}
-	}
-
-	return nil
-}
 
 func deleteVolume(c transaction.TxnCtx) error {
 
@@ -51,7 +28,6 @@ func registerVolDeleteStepFuncs() {
 		name string
 		sf   transaction.StepFunc
 	}{
-		{"vol-delete.DeleteVolfiles", deleteVolfiles},
 		{"vol-delete.Store", deleteVolume},
 	}
 	for _, sf := range sfs {
@@ -87,10 +63,6 @@ func volumeDeleteHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	txn.Steps = []*transaction.Step{
-		{
-			DoFunc: "vol-delete.DeleteVolfiles",
-			Nodes:  volinfo.Nodes(),
-		},
 		{
 			DoFunc: "vol-delete.Store",
 			Nodes:  []uuid.UUID{gdctx.MyUUID},
