--- glusterd2/volgen/templates/brick.graph
@@ -1,19 +1,0 @@
-protocol/server
-performance/decompounder, {{ brick.path }}
-debug/io-stats
-features/quota
-features/index
-features/barrier
-features/marker
-performance/io-threads
-features/upcall
-features/leases
-features/read-only
-features/worm
-features/locks
-features/access-control
-features/bitrot-stub
-features/changelog
-features/changetimerecorder
-features/trash
-storage/posix
