--- glusterd2/volgen/templates/disperse.graph
@@ -1,3 +1,0 @@
-cluster/distribute
-cluster/disperse
-protocol/client
