--- glusterd2/volgen/templates/distreplicate.graph
@@ -1,3 +1,0 @@
-cluster/distribute
-cluster/replicate
-protocol/client
