--- glusterd2/volgen/templates/distribute.graph
@@ -1,2 +1,0 @@
-cluster/dht
-protocol/client
