--- glusterd2/volgen/templates/fuse.graph
@@ -1,10 +1,0 @@
-debug/io-stats
-performance/io-threads
-performance/md-cache
-performance/open-behind
-performance/quick-read
-performance/io-cache
-performance/readdir-ahead
-performance/read-ahead
-performance/write-behind
-cluster.graph
